# swagger.yaml
swagger: "2.0"
info:
  description: "This is the swagger doc for vanDevTimes"
  version: "1.0.0"
  title: "Swagger vanDevTimes"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "vantimes.info@gmail.com"
basePath: "/api"
paths:
  /resources:
    get:
      tags:
        - "Content"
      summary: "get all resources (Resources and feeds)"
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: '#/definitions/ArrayOfPostFull' 
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          schema:
            example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"

    delete:
      tags:
        - "Content"
      summary: "delete content resource with url"
      description: ""
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "delete content resource with url"
          required: true
          schema:
            example: {
             "url": "https://stackoverflow.com/questions/44610613/node-js-mysql-deleting-multiple-rows-with-an-array-of-array"
            }
      responses:
        200:
          description: "successful operation"
          schema:
            example: {
             "url": "https://stackoverflow.com/questions/44610613/node-js-mysql-deleting-multiple-rows-with-an-array-of-array"
            }
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          schema:
            example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
  /resources/pending:
    get:
      tags:
        - "Content"
      summary: "return pending content along with existing reviewers(if any exists)"
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            example: {
              allLinks: [
              { resourceType: "Blog Article",
                categoryName: "Others",
                url: "www.google.com/blog/123",
                title: "Example",
                approvalCount: 2,
                approvalThreshold: 4,
                approver: "John.doe@gmail.com"
              }
              ]
            }
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          schema:
            example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
...
  /resources/approved/user:
    get:
      tags:
        - "Content"
      summary: "return resources that have been approved by users"
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            example: {
              allLinks: [
              { resourceType: "Blog Article",
                categoryName: "Others",
                url: "www.google.com/blog/123",
                title: "Example",
                approvalCount: 2,
                approvalThreshold: 4,
                approver: "John.doe@gmail.com"
              }
              ]
            }
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          schema:
            example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
  /resources/approved:
    get:
      tags:
        - "Content"
      summary: "get all approved resources (Resources and feeds)"
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: '#/definitions/ArrayOfPostFull'
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          schema:
            example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
  /resources/approvedByCategory/:maxNum:
    get:
      tags:
        - "Content"
      summary: "return maxNum of approved content resources from each catagory"
      parameters:
        - name: "maxNum"
          in: "path"
          type: integer
          required: true
          description: "maximum number of approved articles to be retrieved from each category"
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            example: {
              "allApprovedByCategory": [
              {
                "category": "Database",
                "articles": [
                {
                  "link": "https://www.wegalvanize.com/featured/cyber-risk-assessments/1",
                  "title": "How to conduct a cyber-risk assessment",
                  "category": "Database",
                  "approvedTime": "2020-03-30T04:31:26.000Z"
                }
                ]
              },
              {
                "category": "Products",
                "articles": [
                {
                  "link": "https://www.wegalvanize.com/featured/cyber-risk-assessments/4",
                  "title": "How to conduct a cyber-risk assessment",
                  "category": "Products"
                }
                ]
              },
              {
                "category": "Social",
                "articles": [
                {
                  "link": "https://www.wegalvanize.com/featured/cyber-risk-assessments/6",
                  "title": "How to conduct a cyber-risk assessment",
                  "category": "Social"
                }
                ]
              }
              ]
            }
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          schema:
            example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
  /resources/search:
    get:
      tags:
        - "Content"
      summary: "get all resources with matching title and category"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "title search string, category selected from drop down menu in search area (valid category name or 'All')"
          required: true
          schema:
            example: {
              "titleString": "assess",
              "category": "Database"
            }
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ArrayOfPostFull"
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          schema:
            example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
  /resource/{category}:
    get:
      tags:
        - "Content"
      summary: "get all resources by from a specific category"
      parameters:
        -  name: "category"
           in: "path"
           type: string
           required: true
           description: "Category of content resource to return"
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: '#/definitions/ArrayOfPostFull' 
  /resource/approved/{category}:
    get:
      tags:
        - "Content"
      summary: "get all approved resources by from a specific category"
      parameters:
        -  name: "category"
           in: "path"
           type: string
           required: true
           description: "Category of content resource to return"
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: '#/definitions/ArrayOfPostFull'
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          schema:
            example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
  /post/{title}:
    get:
      tags:
        - "Content"
      summary: "get the resource by a specific title"
      parameters:
        -  name: "title"
           in: "path"
           type: string
           required: true
           description: "Title of content resource to return"
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            example: {
              'url': "https://www.google.com"
            }
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          schema:
            example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
  /post:
    post:
      tags:
        - "Content"
      summary: "submit post"
      description: ""
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "url, title, category, and user permission type of a post submission"
          required: true
          schema:
            $ref: '#/definitions/PostSubmission'
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: '#/definitions/PostFull'
        400:
          description: "invalid or dead url"
          schema:
            example: {
              "errorCode": 400,
              "message": "URL entered is invalid/not live."
            }
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          schema:
            example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
  /rss:
    get:
      tags:
        - "Content"
      summary: "get latest articles from a RSS feed"
      description: ""
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "link of the RSS feed"
          required: true
          schema:
            $ref: '#/definitions/RSSFeedSubmission'
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: '#/definitions/ArrayOfResource'
        400:
          description: "invalid or dead url"
          schema:
            example: {
              "errorCode": 400,
              "message": "Feed to poll is not live anymore"
            }
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          schema:
            example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
    post:
      tags:
        - "Content"
      summary: "submit RSS feed"
      description: ""
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "url and user permission type of RSS feed"
          required: true
          schema:
            $ref: '#/definitions/RSSFeedSubmission'
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: '#/definitions/RSSFeedFull'
        400:
          description: "invalid or dead url"
          schema:
            example: {
              "errorCode": 400,
              "message": "URL entered is invalid/not live."
            }
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          schema:
            example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
  /approvals:
    post:
      tags:
        - "Content"
      summary: "approve content resource with url"
      description: ""
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "approver email and url of content to be approved"
          required: true
          schema:
            example: {
              user: "john.doe@gmail.com",
              url: "www.google.com/blog"
            }
      responses:
        200:
          description: "successful operation"
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          schema:
            example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
  /approvals/rss:
    post:
      tags:
        - "Content"
      summary: "approve rss feed with url (poll if approvalCount >= approvalThreshold"
      description: ""
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "approver email, url of content to be approved, and original submitter permissionType"
          required: true
          schema:
            example: {
              "user": "vandevtimesadmuser@gmail.com",
              "url": "https://medium.com/feed/the-launchism",
              "permissionType": "admin"
            }
      responses:
        200:
          description: "successful operation"
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          schema:
            example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
...
  /feed:
    get:
      tags:
        - "Content"
      summary: "site RSS feed"
      description: ""
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation: RSS xml rendered at endpoint path"
        401:
          description: "unauthorized operation"
        500:
          description: "database error"
          schema:
            example: "{Error: ER_BAD_FIELD_ERROR: some detailed information}"
...
definitions:
...
  PostSubmission:
    type: "object"
    required:
      - "url"
      - "title"
      - "category"
      - "permissionType"
    properties:
      url:
        example: "www.google.com"
      title:
        example: "how to make google"
      category:
        example: "front-end"
      permissionType:
        example: "admin"
  RSSFeedSubmission:
    type: "object"
    required:
      - "url"
      - "permissionType"
    properties:
      url:
        example: "https://medium.com/feed/the-launchism"
      permissionType:
        example: "sysManager"
...
  ArrayOfPostFull:
    type: "object"
    required:
      - "resources"
    properties:
      resources:
        example: [
        {
          "resourceType": "Blog Post",
          "approvalThreshold": 3,
          "url": "https://www.wegalvanize.com/featured/cyber-risk-assessments/1",
          "submissionTime": "2020-02-05T08:00:00.000Z",
          "approvedTime": "2020-02-06T08:00:00.000Z",
          "approvalCount": 4,
          "title": "How to conduct a cyber-risk assessment",
          "categoryName": "Database"
        },
        {
          "resourceType": "Blog Post",
          "approvalThreshold": 3,
          "url": "https://www.wegalvanize.com/featured/cyber-risk-assessments/2",
          "submissionTime": "2020-02-05T08:00:00.000Z",
          "approvedTime": "2020-02-06T08:00:00.000Z",
          "approvalCount": 5,
          "title": "How to conduct a cyber-risk assessment",
          "categoryName": "Database"
        }
        ]
...